/ Primary Changes - PDF rendering functionality
Added separate template folder and HTML components for the PDF Document.  While code is reused from the GUI version, the pdf document is rendered separately on an A4 sized page with associated CSS.  The challenge here is to get the page to render properly in A4 when Bootstrap can't interpret the breakpoints (default renders as a mobile page).

PDF rendering is done via an API to api2pdf.  This is not fully implemented.  Some basic persistence is being achieved by pickling data to files.  This is not robust/secure and a data model is required.

Separated client Application into Version 1.0 and Version 2.0.  Version 1.0 is a like for like replacement of the existing App.  Version 2.0 will incorporate design changes.

Some refinements of the loanProjection calculation have been made.  Further refinements are required (and calcs need to be validated).

Some helper functions in Client:Views were introduced to help extract the projection points from the main projection array.

Updated and separated (fully) the css for the client web-app and the other site apps.